Remove ShimmerHost duplicate
