+++ app/src/main/kotlin/it/vfsfitvnm/vimusic/ui/components/ShimmerHost.kt
@@ -3,6 +3,7 @@ package it.vfsfitvnm.vimusic.ui.components
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.graphics.BlendMode
@@ -12,9 +13,14 @@ import androidx.compose.ui.graphics.graphicsLayer
 import com.valentinilk.shimmer.shimmer
 
 @Composable
-fun ShimmerHost(content: @Composable ColumnScope.() -> Unit) {
+fun ShimmerHost(
+    modifier: Modifier = Modifier,
+    horizontalAlignment: Alignment.Horizontal = Alignment.Start,
+    content: @Composable ColumnScope.() -> Unit
+) {
     Column(
-        modifier = Modifier
+        horizontalAlignment = horizontalAlignment,
+        modifier = modifier
             .shimmer()
             .graphicsLayer(alpha = 0.99f)
             .drawWithContent {
